intent-requirements-service/requirements.txt
pydantic>=2.5.0,<3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
colorama>=0.4.6
pyyaml>=6.0
crewai>=0.40.0
crewai-tools>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-core>=0.1.0
mangum>=0.17.0
boto3>=1.34.0